+++ packages/moonstone/Panels/Panel.js
@@ -7,7 +7,7 @@ import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDeco
 import css from './Panel.less';
 
 const spotPanel = (node) => {
-	if (node && !node.contains(document.activeElement)) {
+	if (node && !Spotlight.getCurrent()) {
 		const body = node.querySelector('section .spottable');
 		const header = node.querySelector('header .spottable');
 		const spottable = body || header;