+++ packages/sampler/stories/qa-stories/Spotlight.js
@@ -42,6 +42,11 @@ const style = {
 		margin: '0 12px',
 		padding: '12px'
 	},
+	fittedContainer: {
+		border: '1px dashed blue',
+		margin: '0 12px',
+		padding: '12px'
+	},
 	flexBox: {
 		display: 'flex'
 	},
@@ -247,6 +252,27 @@ storiesOf('Spotlight')
 						<Item onFocus={action('onFocus')} onBlur={action('onBlur')}>1</Item>
 						<Item onFocus={action('onFocus')} onBlur={action('onBlur')}>2</Item>
 						<Item onFocus={action('onFocus')} onBlur={action('onBlur')}>3</Item>
+					</Container>
+				</div>
+			</div>
+		)
+	)
+	.addWithInfo(
+		'Nested Containers',
+		() => (
+			<div>
+				<p>
+					The nested containers below both use a enterTo: &apos;last-focused&apos; configuration.
+					You should be able to naturally 5-way navigate between the items in the containers. Also,
+					attempting to 5-way navigate (left or down) from the application close button should
+					result in the last-focused item being spotted.
+				</p>
+				<div style={style.flexBox}>
+					<Container style={style.fittedContainer} >
+						<Item>Item in a container</Item>
+						<Container style={style.fittedContainer} >
+							<Item>Item in a nested container</Item>
+						</Container>
 					</Container>
 				</div>
 			</div>