+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `moonstone/Panels.Panel` behavior for setting focus after render
+- `moonstone/VirtualList.VirtualGridList` showing empty items when items are continuously added dynamically
+- `moonstone/Picker` to marquee on focus once again
+
 ## [1.0.0-beta.4] - 2017-03-10
 
 ### Added