+++ packages/spotlight/CHANGELOG.md
@@ -2,7 +2,17 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+`spotlight.Spotlight` behavior to properly save the last-focused element for nested containers
+
 ## [1.0.0-beta.4] - 2017-03-10
+
+### Fixed
+
+- `spotlight.Spotlight` pointer behavior where upon immediately entering pointer-mode, hovering over a spottable component may result in the component not receiving focus
 
 ### Changed